fix(form): calendar design doesn't show
