+++ frontend/src/components/AppNavbar.vue
@@ -48,7 +48,9 @@
             d="M15.75 10.5V6a3.75 3.75 0 10-7.5 0v4.5m11.356-1.993l1.263 12c.07.665-.45 1.243-1.119 1.243H4.25a1.125 1.125 0 01-1.12-1.243l1.264-12A1.125 1.125 0 015.513 7.5h12.974c.576 0 1.059.435 1.119 1.007z"
           />
         </svg>
-        <span v-if="cartCount > 0" class="navbar-cart-badge">{{ cartCount > 9 ? '9+' : cartCount }}</span>
+        <span v-if="cartCount > 0" class="navbar-cart-badge">{{
+          cartCount > 9 ? '9+' : cartCount
+        }}</span>
       </button>
     </div>
   </header>
@@ -93,7 +95,11 @@
     font-size: 1.375rem;
     font-weight: 600;
     color: #111827;
-    font-family: 'Instrument Sans', 'DM Sans', -apple-system, sans-serif;
+    font-family:
+      'Instrument Sans',
+      'DM Sans',
+      -apple-system,
+      sans-serif;
     letter-spacing: -0.02em;
     line-height: 1.2;
   }
