+++ frontend/src/css/SubmitAvailability.css
@@ -1,0 +1,387 @@
+/* Submit Availability Page Layout */
+.submit-page {
+  display: flex;
+  justify-content: center;
+  padding: 1rem;
+  height: calc(100vh - 60px);
+  overflow: hidden;
+}
+
+/* Unified Card */
+.submit-card {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 1400px;
+  height: 100%;
+  background: var(--bg-card);
+  border: 1px solid var(--border-subtle);
+  border-radius: 12px;
+  box-shadow: var(--shadow-card);
+  overflow: hidden;
+}
+
+/* Toolbar */
+.submit-toolbar {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  padding: 0.875rem 1rem;
+  border-bottom: 1px solid var(--border-subtle);
+  flex-shrink: 0;
+}
+
+.toolbar-section {
+  display: flex;
+  align-items: center;
+}
+
+.toolbar-section--teacher {
+  flex: 1;
+  min-width: 0;
+}
+
+.toolbar-section--counter {
+  gap: 0.75rem;
+  padding: 0 1rem;
+  border-left: 1px solid var(--border-subtle);
+  border-right: 1px solid var(--border-subtle);
+}
+
+.toolbar-section--action {
+  flex-shrink: 0;
+}
+
+/* Teacher Select */
+.teacher-select {
+  width: 100%;
+  max-width: 280px;
+  padding: 0.625rem 0.875rem;
+  font-size: 0.875rem;
+  font-family: 'IBM Plex Sans', sans-serif;
+  color: var(--text-primary);
+  background: var(--bg-subtle);
+  border: 1px solid var(--border-medium);
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  appearance: none;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='none' stroke='%23585a63' viewBox='0 0 24 24'%3E%3Cpath stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M8 10l4 4 4-4'/%3E%3C/svg%3E");
+  background-repeat: no-repeat;
+  background-position: right 0.75rem center;
+  background-size: 1rem;
+  padding-right: 2.25rem;
+}
+
+.teacher-select:hover {
+  border-color: var(--border-strong);
+  background-color: var(--bg-card);
+}
+
+.teacher-select:focus {
+  outline: none;
+  border-color: var(--primary-indigo);
+  box-shadow: 0 0 0 3px rgba(62, 76, 122, 0.1);
+}
+
+/* Counter */
+.counter-label {
+  font-family: 'JetBrains Mono', monospace;
+  font-size: 0.6875rem;
+  font-weight: 600;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  color: var(--text-secondary);
+  white-space: nowrap;
+}
+
+.counter-value {
+  font-family: 'JetBrains Mono', monospace;
+  font-size: 1.125rem;
+  font-weight: 600;
+  color: var(--primary-indigo);
+  min-width: 1.5ch;
+  text-align: center;
+}
+
+/* Calendar Container */
+.submit-calendar-container {
+  position: relative;
+  flex: 1;
+  min-height: 0;
+  padding: 0.75rem;
+  overflow: hidden;
+}
+
+.submit-calendar {
+  height: 100%;
+}
+
+/* Toast Notifications */
+.submit-toast {
+  position: fixed;
+  bottom: 2rem;
+  right: 2rem;
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  padding: 1rem 1.5rem;
+  border-radius: 8px;
+  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.12);
+  animation: submit-toast-in 0.4s cubic-bezier(0.34, 1.56, 0.64, 1);
+  z-index: 100;
+  font-size: 0.875rem;
+  font-family: 'IBM Plex Sans', sans-serif;
+  font-weight: 500;
+}
+
+.submit-toast--success {
+  background: var(--accent-sage);
+  color: white;
+}
+
+.submit-toast--error {
+  background: var(--accent-coral);
+  color: white;
+}
+
+.submit-toast-icon {
+  width: 1.25rem;
+  height: 1.25rem;
+  flex-shrink: 0;
+}
+
+.submit-toast-close {
+  margin-left: 0.75rem;
+  color: rgba(255, 255, 255, 0.8);
+  background: transparent;
+  border: none;
+  font-size: 1.25rem;
+  cursor: pointer;
+  width: 1.5rem;
+  height: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 4px;
+  transition: all 0.2s ease;
+}
+
+.submit-toast-close:hover {
+  background: rgba(255, 255, 255, 0.15);
+}
+
+/* Confirm Dialog */
+.submit-confirm-overlay {
+  position: fixed;
+  inset: 0;
+  background: rgba(26, 28, 35, 0.6);
+  backdrop-filter: blur(4px);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 200;
+  animation: submit-fade-in 0.2s ease-out;
+}
+
+.submit-confirm-dialog {
+  background: var(--bg-card);
+  border: 1px solid var(--border-medium);
+  border-top: 3px solid var(--accent-sage);
+  border-radius: 12px;
+  padding: 1.75rem;
+  max-width: 420px;
+  margin: 0 1rem;
+  animation: submit-dialog-in 0.3s cubic-bezier(0.34, 1.56, 0.64, 1);
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.12);
+}
+
+.submit-confirm-title {
+  font-size: 1.125rem;
+  font-weight: 600;
+  font-family: 'Instrument Sans', sans-serif;
+  color: var(--text-primary);
+  margin: 0 0 1.25rem 0;
+}
+
+.submit-confirm-content {
+  display: flex;
+  flex-direction: column;
+  gap: 0.875rem;
+  margin-bottom: 1.5rem;
+}
+
+.submit-confirm-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.5rem 0;
+}
+
+.submit-confirm-label {
+  font-family: 'JetBrains Mono', monospace;
+  font-size: 0.6875rem;
+  font-weight: 600;
+  letter-spacing: 0.1em;
+  color: var(--text-secondary);
+  text-transform: uppercase;
+}
+
+.submit-confirm-value {
+  font-family: 'IBM Plex Sans', sans-serif;
+  font-size: 0.9375rem;
+  font-weight: 500;
+  color: var(--text-primary);
+}
+
+.submit-confirm-slots {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding: 1rem;
+  background: var(--bg-subtle);
+  border-radius: 8px;
+  border: 1px solid var(--border-subtle);
+}
+
+.submit-confirm-slot {
+  font-family: 'JetBrains Mono', monospace;
+  font-size: 0.8125rem;
+  color: var(--text-primary);
+  padding: 0.25rem 0;
+}
+
+.submit-confirm-actions {
+  display: flex;
+  gap: 0.875rem;
+  justify-content: flex-end;
+}
+
+.submit-confirm-btn {
+  padding: 0.75rem 1.25rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  font-family: 'IBM Plex Sans', sans-serif;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.submit-confirm-btn--cancel {
+  color: var(--text-primary);
+  background: var(--bg-subtle);
+  border: 1px solid var(--border-medium);
+}
+
+.submit-confirm-btn--cancel:hover {
+  background: var(--border-subtle);
+}
+
+.submit-confirm-btn--confirm {
+  color: white;
+  background: var(--accent-sage);
+  border: none;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.submit-confirm-btn--confirm:hover {
+  filter: brightness(0.9);
+}
+
+/* Animations */
+@keyframes submit-toast-in {
+  from {
+    opacity: 0;
+    transform: translateY(16px) scale(0.95);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+}
+
+@keyframes submit-dialog-in {
+  from {
+    opacity: 0;
+    transform: scale(0.95) translateY(8px);
+  }
+  to {
+    opacity: 1;
+    transform: scale(1) translateY(0);
+  }
+}
+
+@keyframes submit-fade-in {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+/* Mobile Responsive */
+@media (max-width: 768px) {
+  .submit-page {
+    padding: 0.75rem;
+  }
+
+  .submit-title {
+    font-size: 1.25rem;
+  }
+
+  .submit-toolbar {
+    flex-wrap: wrap;
+    padding: 0.75rem;
+    gap: 0.75rem;
+  }
+
+  .toolbar-section--teacher {
+    width: 100%;
+    max-width: 100%;
+  }
+
+  .teacher-select {
+    max-width: 100%;
+  }
+
+  .toolbar-section--counter {
+    padding-left: 0.75rem;
+    padding-right: 0.75rem;
+  }
+
+  .submit-toast {
+    right: 1rem;
+    left: 1rem;
+    bottom: 1rem;
+  }
+
+  .submit-confirm-dialog {
+    padding: 1.5rem;
+    margin: 0 1rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .submit-toolbar {
+    flex-direction: column;
+    align-items: stretch;
+  }
+
+  .toolbar-section--counter {
+    border-left: none;
+    border-right: none;
+    border-top: 1px solid var(--border-subtle);
+    border-bottom: 1px solid var(--border-subtle);
+    padding: 0.75rem;
+    justify-content: space-between;
+  }
+
+  .toolbar-section--action {
+    width: 100%;
+  }
+
+  .toolbar-section--action :deep(button) {
+    width: 100%;
+  }
+}
